Removes Codenames game client and server implementation
Eliminates all source code for both the client and server, effectively decommissioning the Codenames application from the codebase.
Simplifies project structure and reduces maintenance overhead for unused or obsolete components.
